python/youtube-rtmp-streamer/pyproject.toml
[project]
name = "avatartalk-youtube-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "asgiref>=3.9.1",
    "google-api-python-client>=2.184.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "livekit>=1.0.12",
    "livekit-api>=1.0.5",
    "nltk>=3.9.2",
    "openai>=1.102.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "websockets>=15.0.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "ruff>=0.12.9",
]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    #### modules
    "ANN", # flake8-annotations
    "COM", # flake8-commas
    "C90", # mccabe complexity
    "DJ",  # django
    "EXE", # flake8-executable
    "T10", # debugger
    "TID", # flake8-tidy-imports
    "T201", # prints

    #### specific rules
    "D100",   # ignore missing docs
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    "D200",
    "D205",
    "D212",
    "D400",
    "D401",
    "D415",
    "E402",   # false positives for local imports
    "E501",   # line too long
    "G004",
    "TRY003", # external messages in exceptions are too verbose
    "TD002",
    "TD003",
    "FIX002", # too verbose descriptions of todos
    "TRY300",
    "TRY301",
    "LOG015",
    "PTH110",
    "PTH123",
    "PLR2004",
    "S104",
    "S311",
    "BLE001",
    "RUF",
    "ARG",
    "PLR0913",
    "PLR0915",
    "TRY401",
    "EM101"
]
